requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
beautifulsoup4==4.11.1
    # via -r requirements.in
build==0.10.0
    # via pip-tools
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via pip-tools
django==4.0.5
    # via
    #   -r requirements.in
    #   djangorestframework
django-cleanup==7.0.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
packaging==23.1
    # via build
pip-tools==6.13.0
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pytz==2022.1
    # via djangorestframework
requests==2.28.1
    # via -r requirements.in
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via build
urllib3==1.26.11
    # via requests
wheel==0.40.0
    # via pip-tools
whitenoise==6.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
